test: move consumer e2e to sperated spec file
